WIP: some changes to pom.xml files
