+++ pom.xml
@@ -124,6 +124,8 @@
 		<jackson-core.version>2.15.2</jackson-core.version>
 		<jackson-databind.version>2.15.2</jackson-databind.version>
 		<jackson-dataformat-yaml.version>2.15.2</jackson-dataformat-yaml.version>
+		<jackson-datatype.version>2.15.2</jackson-datatype.version>
+		<jakarta-annotation.version>1.3.5</jakarta-annotation.version>
 		<jama.version>1.0.3</jama.version>
 		<jaxb-api.version>2.3.0</jaxb-api.version>
 		<jaxb-core.version>2.3.0</jaxb-core.version>
