+++ vcell-restclient/pom.xml
@@ -199,27 +199,32 @@
 <!--    </profiles>-->
 
     <dependencies>
+        <dependency>
+            <groupId>org.vcell</groupId>
+            <artifactId>vcell-util</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
 
         <!-- JSON processing: jackson -->
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-core</artifactId>
-            <version>${jackson-version}</version>
+            <version>${jackson-core.version}</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-annotations</artifactId>
-            <version>${jackson-version}</version>
+            <version>${jackson-annotations.version}</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-            <version>${jackson-version}</version>
+            <version>${jackson-databind.version}</version>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.datatype</groupId>
             <artifactId>jackson-datatype-jsr310</artifactId>
-            <version>${jackson-version}</version>
+            <version>${jackson-datatype.version}</version>
         </dependency>
         <dependency>
             <groupId>com.nimbusds</groupId>
@@ -242,7 +247,7 @@
         <dependency>
             <groupId>jakarta.annotation</groupId>
             <artifactId>jakarta.annotation-api</artifactId>
-            <version>${jakarta-annotation-version}</version>
+            <version>${jakarta-annotation.version}</version>
             <scope>provided</scope>
         </dependency>
 
@@ -274,16 +279,6 @@
             <scope>compile</scope>
         </dependency>
 
-
     </dependencies>
 
-    <properties>
-<!--        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>-->
-<!--        <maven.compiler.source>11</maven.compiler.source>-->
-<!--        <maven.compiler.target>11</maven.compiler.target>-->
-        <jackson-version>2.15.2</jackson-version>
-<!--        <jackson-databind-nullable-version>0.2.6</jackson-databind-nullable-version>-->
-        <jakarta-annotation-version>1.3.5</jakarta-annotation-version>
-<!--        <spotless.version>2.27.2</spotless.version>-->
-    </properties>
 </project>
